+++ app/Modules/Integration/Config/config.php
@@ -1,0 +1,10 @@
+<?php
+
+return [
+    'name' => 'Integration',
+    'hrms' => [
+        'host'          => env('HRMS_BASE_LINK', 'http://hrms.test/'),
+        'client_id'     => env('HRMS_CLIENT_ID', ''),
+        'client_secret' => env('HRMS_CLIENT_SECRET', ''),
+    ],
+];