+++ app/Modules/Shifts/Http/Controllers/ScheduleController.php
@@ -2,11 +2,14 @@
 
 namespace App\Modules\Shifts\Http\Controllers;
 
+use App\Modules\Integration\Jobs\HRMS\DeleteScheduleIntegrationJob;
+use App\Modules\Integration\Jobs\HRMS\EditScheduleIntegrationJob;
 use Carbon\Carbon;
 use Faker\Factory;
 use App\Modules\Teams\Services\TeamService;
 use App\Modules\Shifts\Services\MemberScheduleService;
 use App\Modules\Core\Http\Controllers\AbstractCoreController;
+use App\Modules\Integration\Jobs\HRMS\CreateScheduleIntegrationJob;
 use App\Modules\Shifts\Http\Requests\MemberSchedules\CreateMemberScheduleRequest;
 use App\Modules\Shifts\Http\Requests\MemberSchedules\DeleteMemberScheduleRequest;
 use App\Modules\Shifts\Http\Requests\MemberSchedules\UpdateMemberScheduleRequest;
@@ -49,6 +52,8 @@ class ScheduleController extends AbstractCoreController
         $memberScheduleAdded = $this->memberScheduleService->create($request->validated());
 
         if ($memberScheduleAdded) {
+            CreateScheduleIntegrationJob::dispatch($memberScheduleAdded);
+
             return response()->json([
                 'message'    => 'success',
                 'scheduleId' => $memberScheduleAdded->id,
@@ -80,6 +85,10 @@ class ScheduleController extends AbstractCoreController
         }
 
         if ($this->memberScheduleService->update($changes, $memberScheduleId)) {
+            $editedSchedule = $this->memberScheduleService->read($memberScheduleId);
+
+            EditScheduleIntegrationJob::dispatch($editedSchedule);
+
             return response()->json([
                 'message'  => 'success',
             ]);
@@ -91,6 +100,7 @@ class ScheduleController extends AbstractCoreController
     public function deleteSchedule(DeleteMemberScheduleRequest $request)
     {
         if ($this->memberScheduleService->delete($request->validated()['id'])) {
+            DeleteScheduleIntegrationJob::dispatch($request->validated()['id']);
             return response()->json([
                 'message'  => 'success',
             ]);