+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [5.0.1] - 2023-02-18
+### Added
+- Add integration with the HRMS when create shift / schedule
+- Add integration with the HRMS when edit shift / schedule
+- Add integration with the HRMS when delete shift / schedule
+
+
 ## [4.0.3] - 2022-08-23
 ### Added
 - Add logs viewer for tracing